chore: rename lib to utils
